fix bug input double number

+++ app/src/main/java/com/example/itemnote/data/model/ShopModel.kt
@@ -8,5 +8,5 @@ data class ShopModel(
     val name: String = "",
     val date: String = "",
     val location: String = "",
-    val price: Int = 0
+    val price: Double = 0.0
 )
